requirements.txt
Django==2.0.4

pytz==2018.4

djangorestframework
django-ckeditor
django-cors-headers
Pillow